refactor(native): extract hotkeys from keyboard (#2831)
Move handle_common_hotkey, handle_ctrl_hotkey, adjust_volume,
handle_cartridge_switch_key, and key_code_to_filter_char into
keyboard/hotkeys.rs. Cross-module fns are pub(super); mod.rs and the
console dispatch delegate via hotkeys::. No behavior change.
